+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 All notable changes to the "Go Current - Workspace" extension will be documented in this file.
 
+## [0.2.0] - 2019-09-16
+### Changed
+- Adapted to new Go Current version.
+
 ## [0.1.0]
 - User can remove installed package groups trough *Remove packages* command.
-- User prompt to update launch.json when nav-server package installed.
+- User prompt to update launch.json when bc-server package installed.
 - User prompt with json document with arguments to fill when installing a package requiring arguments.
 - Install and update package groups, defined on project level.
 - Initial release